Exit from port-forward if an error occurs
This should be helpful in situations where a pod's container closes and
reopens its socket, causing the port-forward to get stuck in a state
where it's no longer able to communicate with the pod.  Exiting and
restarting the port-forward restores communication.

To benefit from this change, you probably should run "kubectl
port-forward" in a loop.  For example:

  while true; do kubectl port-forward --address 0.0.0.0 deployment/my-app 8080:5001; done